+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   * Added 'visualize_multiple_sensors' example
   * Added 'check_lidar_bb' util script
   * Improved manual_control: now cameras are set in relation with car size
+  * Added CHRONO library for vehicle dynamics simulation (https://projectchrono.org/)
   * Added performance benchmarking section to documentation
   * CARLA is compatible with the last RoadRunner nomenclature for road assets
   * Fixed a bug when importing a FBX map with some **_** in the FBX name